+++ src/Chunkyard.Tests/Core/FuzzyTests.cs
@@ -1,4 +1,5 @@
-﻿using Chunkyard.Core;
+using System;
+using Chunkyard.Core;
 using Xunit;
 
 namespace Chunkyard.Tests.Core
@@ -6,20 +7,37 @@ namespace Chunkyard.Tests.Core
     public static class FuzzyTests
     {
         [Fact]
-        public static void EmptyPattern_Matches_Everything()
+        public static void Empty_Pattern_Matches_Everything()
         {
-            var fuzzy = new Fuzzy("");
+            var fuzzy1 = Fuzzy.MatchAll;
+            var fuzzy2 = new Fuzzy(new[] { "" }, true);
+            var fuzzy3 = new Fuzzy(new[] { "" }, false);
 
-            Assert.True(fuzzy.IsMatch("some text!"));
+            Assert.True(fuzzy1.IsMatch("some text!"));
+            Assert.True(fuzzy2.IsMatch("some text!"));
+            Assert.True(fuzzy3.IsMatch("some text!"));
+        }
+
+        [Fact]
+        public static void No_Parameter_Matches_Maybe()
+        {
+            var fuzzy1 = new Fuzzy(Array.Empty<string>(), true);
+            var fuzzy2 = new Fuzzy(Array.Empty<string>(), false);
+            var fuzzy3 = Fuzzy.MatchNothing;
+
+            Assert.True(fuzzy1.IsMatch("some text!"));
+            Assert.False(fuzzy2.IsMatch("some text!"));
+            Assert.False(fuzzy3.IsMatch("some text!"));
         }
 
         [Fact]
         public static void Spaces_Are_Treated_As_Wildcards()
         {
-            var fuzzy = new Fuzzy("He ld");
+            var fuzzy = new Fuzzy(new[] { "He ld", "HE LD" }, true);
 
             Assert.True(fuzzy.IsMatch("Hello World!"));
             Assert.True(fuzzy.IsMatch("Held"));
+            Assert.True(fuzzy.IsMatch("HELLO WORLD!"));
             Assert.False(fuzzy.IsMatch("hello world!"));
             Assert.False(fuzzy.IsMatch("Goodbye World!"));
         }