+++ src/Chunkyard/Cli/CheckOptions.cs
@@ -1,4 +1,5 @@
-﻿using CommandLine;
+﻿using System.Collections.Generic;
+using CommandLine;
 
 namespace Chunkyard.Cli
 {
@@ -8,12 +9,12 @@ namespace Chunkyard.Cli
         public CheckOptions(
             string repository,
             int logPosition,
-            string includeFuzzy,
+            IEnumerable<string> includePatterns,
             bool shallow)
         {
             Repository = repository;
             LogPosition = logPosition;
-            IncludeFuzzy = includeFuzzy;
+            IncludePatterns = includePatterns;
             Shallow = shallow;
         }
 
@@ -23,8 +24,8 @@ namespace Chunkyard.Cli
         [Option('s', "snapshot", Required = false, HelpText = "The snapshot ID", Default = Commands.LatestLogPosition)]
         public int LogPosition { get; }
 
-        [Option('i', "include", Required = false, HelpText = "The fuzzy pattern for files to include")]
-        public string IncludeFuzzy { get; }
+        [Option('i', "include", Required = false, HelpText = "The fuzzy patterns for files to include")]
+        public IEnumerable<string> IncludePatterns { get; }
 
         [Option("shallow", Required = false, HelpText = "Check if chunks exist without further validation", Default = false)]
         public bool Shallow { get; }