+++ src/Chunkyard/Cli/ShowOptions.cs
@@ -1,4 +1,5 @@
-﻿using CommandLine;
+﻿using System.Collections.Generic;
+using CommandLine;
 
 namespace Chunkyard.Cli
 {
@@ -8,11 +9,11 @@ namespace Chunkyard.Cli
         public ShowOptions(
             string repository,
             int logPosition,
-            string includeFuzzy)
+            IEnumerable<string> includePatterns)
         {
             Repository = repository;
             LogPosition = logPosition;
-            IncludeFuzzy = includeFuzzy;
+            IncludePatterns = includePatterns;
         }
 
         [Option('r', "repository", Required = true, HelpText = "The repository path")]
@@ -21,7 +22,7 @@ namespace Chunkyard.Cli
         [Option('s', "snapshot", Required = false, HelpText = "The snapshot ID", Default = Commands.LatestLogPosition)]
         public int LogPosition { get; }
 
-        [Option('i', "include", Required = false, HelpText = "The fuzzy pattern for files to include")]
-        public string IncludeFuzzy { get; }
+        [Option('i', "include", Required = false, HelpText = "The fuzzy patterns for files to include")]
+        public IEnumerable<string> IncludePatterns { get; }
     }
 }